+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+## [0.10.1] - 2024-03-12
+
+### Added
+
+- Updated `type-fest` dependency version to `4.12.0`
+
+## [0.10.0] - 2024-03-11
+
 ### Fixed
 
 - Json serializer control character escape fixed. Some unicode characters
@@ -29,6 +37,12 @@
 - The event handler registration order changed. Now the handlers are called in the opposite order of the registration.
 - Added the padded rectangle, the bounding rectangle and the align parameter to the draw text event object.
 - Tooltip works on marker labels too.
+- New plugins and plugin hooks introduced: 
+  - plugin: scheduler - plugin resposible for scheduling the rendering
+  - plugin: clock - supplying the current time for the animation
+  - plugin: canvasRenderer - plugin for rendering the chart on a htmlcanvas compatible canvas
+  - hook: start - hook for starting the rendering loop
+  - hook: render - hook for rendering the chart 
 
 ## [0.9.3] - 2023-12-20